+++ gui/ports/macOS/build.sh
@@ -98,10 +98,15 @@ $pv/bin/python3 setup.py py2app &&
 # echo "-- Archiving ----------------------------------------------------------------------------------------------------" &&
 # tar cvf - -C dist gui-warp.app | gzip --best > gui-warp.app.tgz &&
 
+# Installing the app-launcher - starter
+mv dist/gui-warp.app/Contents/MacOS/gui-warp dist/gui-warp.app/Contents/MacOS/app
+cp starter dist/gui-warp.app/Contents/MacOS/gui-warp
+
 echo "-- Creating DMG -------------------------------------------------------------------------------------------------" &&
 mkdir GUI-Warp
 mv dist/gui-warp.app GUI-Warp
 ln -s /Applications GUI-Warp
+
 hdiutil create GUI-Warp-tmp.dmg -ov -volname "GUI-Warp" -fs HFS+ -srcfolder "GUI-Warp"
 hdiutil convert GUI-Warp-tmp.dmg -format UDZO -o GUI-Warp.dmg